restructured the project slightly, fixed an issue where email invite goes, even if email is already registered
